requirements.txt
PySide6>=6.6.0
mysql-connector-python>=8.2.0
python-dotenv>=1.0.0
psutil>=5.9.0